# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateJetStreamRequest:
    properties:
      allow_direct:
        example: false
        type: boolean
      allow_rollup_hdrs:
        example: false
        type: boolean
      cluster_id:
        example: cluster-123
        type: string
      compression:
        allOf:
        - $ref: '#/definitions/models.JetStreamCompressionType'
        example: none
      deny_delete:
        example: false
        type: boolean
      deny_purge:
        example: false
        type: boolean
      description:
        example: Event logging stream
        type: string
      discard:
        allOf:
        - $ref: '#/definitions/models.JetStreamDiscardPolicy'
        example: old
      duplicate_window:
        description: 2分钟（秒）
        example: 120
        type: integer
      max_age:
        description: 24小时（秒）
        example: 86400
        type: integer
      max_bytes_unit:
        allOf:
        - $ref: '#/definitions/models.StorageUnit'
        description: 最大字节数单位
        example: GB
      max_bytes_value:
        description: 最大字节数值
        example: 1
        type: number
      max_consumers:
        example: 100
        type: integer
      max_msg_size_unit:
        allOf:
        - $ref: '#/definitions/models.StorageUnit'
        description: 最大单条消息大小单位
        example: MB
      max_msg_size_value:
        description: 最大单条消息大小数值
        example: 1
        type: number
      max_msgs:
        description: 限制配置 - 带单位的字节数
        example: 1000000
        type: integer
      metadata:
        $ref: '#/definitions/models.MetadataMap'
      mirror_direct:
        example: false
        type: boolean
      name:
        example: EVENTS
        type: string
      nats_operate_user_id:
        example: user-123
        type: string
      no_ack:
        example: false
        type: boolean
      placement_cluster:
        description: 放置配置
        type: string
      placement_tags:
        items:
          type: string
        type: array
      replicas:
        description: 副本和持久化配置
        example: 1
        type: integer
      retention:
        allOf:
        - $ref: '#/definitions/models.JetStreamRetentionPolicy'
        example: limits
      storage:
        allOf:
        - $ref: '#/definitions/models.JetStreamStorageType'
        example: file
      subjects:
        description: 流配置
        example:
        - events.*
        - logs.*
        items:
          type: string
        type: array
    required:
    - cluster_id
    - name
    - nats_operate_user_id
    - subjects
    type: object
...
  models.CustomTime:
    properties:
      time.Time:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.JetStream:
    properties:
      allow_direct:
        type: boolean
      allow_rollup_hdrs:
        type: boolean
      bytes_unit:
        allOf:
        - $ref: '#/definitions/models.StorageUnit'
        description: 当前字节数单位
      bytes_value:
        description: 当前字节数值
        type: number
      cluster:
        $ref: '#/definitions/models.Cluster'
      cluster_id:
        type: string
      compression:
        $ref: '#/definitions/models.JetStreamCompressionType'
      created_at:
        allOf:
        - $ref: '#/definitions/models.CustomTime'
        description: 时间戳
      deny_delete:
        type: boolean
      deny_purge:
        type: boolean
      description:
        type: string
      discard:
        $ref: '#/definitions/models.JetStreamDiscardPolicy'
      discard_new_per_subject:
        description: 高级配置
        type: boolean
      duplicate_window:
        description: 重复检测窗口（秒）
        type: integer
      first_seq:
        description: 第一个序列号
        type: integer
      id:
        type: string
      last_seq:
        description: 最后一个序列号
        type: integer
      max_age:
        description: 消息最大存活时间（秒）
        type: integer
      max_bytes_unit:
        allOf:
        - $ref: '#/definitions/models.StorageUnit'
        description: 最大字节数单位
      max_bytes_value:
        description: 最大字节数值
        type: number
      max_consumers:
        description: 最大消费者数量
        type: integer
      max_msg_size_unit:
        allOf:
        - $ref: '#/definitions/models.StorageUnit'
        description: 最大单条消息大小单位
      max_msg_size_value:
        description: 最大单条消息大小数值
        type: number
      max_msgs:
        description: 限制配置
        type: integer
      max_msgs_per_subject:
        description: 每个主题的最大消息数量
        type: integer
      messages:
        description: 统计信息
        type: integer
      metadata:
        $ref: '#/definitions/models.MetadataMap'
      mirror_direct:
        type: boolean
      name:
        type: string
      nats_operate_user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联的NATS操作用户信息（不存储在数据库中）
      nats_operate_user_id:
        description: 重构字段：操作NATS的用户身份
        type: string
      no_ack:
        description: 是否禁用确认
        type: boolean
      num_consumers:
        description: 当前消费者数量
        type: integer
      placement_cluster:
        description: 放置标签和元数据
        type: string
      placement_tags:
        items:
          type: string
        type: array
      replicas:
        description: 副本和持久化配置
        type: integer
      retention:
        $ref: '#/definitions/models.JetStreamRetentionPolicy'
      sealed:
        description: 是否密封流
        type: boolean
      status:
        description: active, inactive, error
        type: string
      storage:
        $ref: '#/definitions/models.JetStreamStorageType'
      subjects:
        description: 流配置
        items:
          type: string
        type: array
      sync_message:
        type: string
      sync_status:
        allOf:
        - $ref: '#/definitions/models.JetStreamSyncStatus'
        description: 新增：同步状态管理
      updated_at:
        $ref: '#/definitions/models.CustomTime'
    type: object
...
  models.JetStreamCompressionType:
    enum:
    - none
    - s2
    type: string
    x-enum-varnames:
    - JetStreamCompressionNone
    - JetStreamCompressionS2
...
  models.JetStreamConfigDiff:
    properties:
      cluster_config:
        $ref: '#/definitions/models.StreamConfig'
      database_config:
        $ref: '#/definitions/models.StreamConfig'
      differences:
        items:
          $ref: '#/definitions/models.ConfigDifference'
        type: array
      has_difference:
        type: boolean
    type: object
...
  models.JetStreamResponse:
    properties:
      account_details:
        items:
          $ref: '#/definitions/models.AccountDetail'
        type: array
      accounts:
        type: integer
      api:
        $ref: '#/definitions/models.APIStats'
      bytes:
        type: integer
      config:
        $ref: '#/definitions/models.Config'
      consumers:
        type: integer
      ha_assets:
        type: integer
      limits:
        additionalProperties: true
        type: object
      memory_unit:
        $ref: '#/definitions/models.StorageUnit'
      memory_value:
        type: number
      messages:
        type: integer
      meta_cluster:
        $ref: '#/definitions/models.MetaCluster'
      now:
        type: string
      reserved_memory_unit:
        $ref: '#/definitions/models.StorageUnit'
      reserved_memory_value:
        type: number
      reserved_storage_unit:
        $ref: '#/definitions/models.StorageUnit'
      reserved_storage_value:
        type: number
      server_id:
        type: string
      storage_unit:
        $ref: '#/definitions/models.StorageUnit'
      storage_value:
        type: number
      streams:
        type: integer
      total:
        type: integer
    type: object
...
  models.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.UpdateJetStreamRequest:
    properties:
      allow_direct:
        type: boolean
      allow_rollup_hdrs:
        type: boolean
      deny_delete:
        type: boolean
      deny_purge:
        type: boolean
      description:
        type: string
      discard:
        $ref: '#/definitions/models.JetStreamDiscardPolicy'
      duplicate_window:
        type: integer
      max_age:
        type: integer
      max_bytes_unit:
        allOf:
        - $ref: '#/definitions/models.StorageUnit'
        description: 最大字节数单位
      max_bytes_value:
        description: 最大字节数值
        type: number
      max_consumers:
        type: integer
      max_msg_size_unit:
        allOf:
        - $ref: '#/definitions/models.StorageUnit'
        description: 最大单条消息大小单位
      max_msg_size_value:
        description: 最大单条消息大小数值
        type: number
      max_msgs:
        description: 限制配置 - 带单位的字节数
        type: integer
      metadata:
        allOf:
        - $ref: '#/definitions/models.MetadataMap'
        description: 元数据
      mirror_direct:
        type: boolean
      nats_operate_user_id:
        example: user-123
        type: string
      no_ack:
        type: boolean
      placement_tags:
        description: 放置配置
        items:
          type: string
        type: array
      replicas:
        description: 副本和持久化配置
        type: integer
      retention:
        $ref: '#/definitions/models.JetStreamRetentionPolicy'
      subjects:
        description: 流配置（某些字段创建后不能修改）
        items:
          type: string
        type: array
    required:
    - nats_operate_user_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing NATS control plane with multi-cluster support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NATS Control API
  version: "1.0"
paths:
...
  /jetstreams:
    get:
      consumes:
      - application/json
      description: Retrieve a list of JetStreams with optional filtering and pagination
      parameters:
      - description: Filter by NATS operate user ID
        in: query
        name: nats_operate_user_id
        type: string
      - description: Filter by status (active, inactive, error)
        in: query
        name: status
        type: string
      - description: Filter by sync status (pending, synced, failed)
        in: query
        name: sync_status
        type: string
      - description: Search by name or description
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: Sort field (created_at, name)
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.JetStream'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List JetStreams
      tags:
      - JetStream
    post:
      consumes:
      - application/json
      description: Create a new JetStream on the specified NATS cluster
      parameters:
      - description: JetStream configuration
        in: body
        name: jetstream
        required: true
        schema:
          $ref: '#/definitions/models.CreateJetStreamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.JetStream'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new JetStream
      tags:
      - JetStream
  /jetstreams/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a JetStream from the cluster (not from database)
      parameters:
      - description: JetStream ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete a JetStream
      tags:
      - JetStream
    get:
      consumes:
      - application/json
      description: Retrieve a specific JetStream by its ID, including real-time statistics
      parameters:
      - description: JetStream ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JetStreamResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get a JetStream
      tags:
      - JetStream
    put:
      consumes:
      - application/json
      description: Update an existing JetStream configuration
      parameters:
      - description: JetStream ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated JetStream configuration
        in: body
        name: jetstream
        required: true
        schema:
          $ref: '#/definitions/models.UpdateJetStreamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JetStream'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update a JetStream
      tags:
      - JetStream
  /jetstreams/{id}/diff:
    get:
      consumes:
      - application/json
      description: Compare database configuration with cluster configuration and return
        differences
      parameters:
      - description: JetStream ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JetStreamConfigDiff'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Compare JetStream configuration
      tags:
      - JetStream
  /jetstreams/{id}/retry:
    post:
      consumes:
      - application/json
      description: Retry creating a JetStream on NATS that previously failed but was
        saved to database
      parameters:
      - description: JetStream ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JetStream'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Retry JetStream creation
      tags:
      - JetStream
  /jetstreams/batch-delete:
    post:
      consumes:
      - application/json
      description: Delete multiple JetStreams in a single request
      parameters:
      - description: Array of JetStream IDs to delete
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Batch delete JetStreams
      tags:
      - JetStream
  /jetstreams/stats:
    get:
      consumes:
      - application/json
      description: Retrieve aggregated statistics for all JetStreams
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get JetStream statistics
      tags:
      - JetStream
  /jetstreams/validate-name:
    get:
      consumes:
      - application/json
      description: Check if a JetStream name is available for a specific user
      parameters:
      - description: NATS Operate User ID
        in: query
        name: nats_operate_user_id
        required: true
        type: string
      - description: JetStream name to validate
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Validate JetStream name
      tags:
      - JetStream
...
  /users/{id}/jetstreams:
    get:
      consumes:
      - application/json
      description: Retrieve all JetStreams created by a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.JetStream'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get JetStreams by user
      tags:
      - JetStream
...
schemes:
- http
- https
swagger: "2.0"